Changed Laser Calculations
Previously the laser had n bounces, and each segment had a range of n.

Now there is a max range, and max bounces. Each bounce subtracts from the range, based on how long the segment was. The laser either stops at its max range, or after it has hit its maxBounces. It is more intuitive to determine the lasers length.